% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_package_functions.R
\name{alt_form}
\alias{alt_form}
\title{Formula for damped oscillator with phase and equilibrium shift.}
\usage{
alt_form(a, gam, omega, phi, y_shift, t)
}
\arguments{
\item{a}{Amplitude}

\item{gam}{Forcing coefficient (amount of damping/driving)}

\item{omega}{Radial frequency}

\item{phi}{Phase Shift (radians)}

\item{y_shift}{Equilibrium shift}

\item{t}{time}
}
\value{
result of inputs into formula
}
\description{
Formula for damped oscillator with phase and equilibrium shift.
}
