% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_package_functions.R
\name{avg_all_rep}
\alias{avg_all_rep}
\title{Function to calculate the average of replicates for all given expressions. Used primarily for cases with multiple replicates.}
\usage{
avg_all_rep(num_reps, genes, timen)
}
\arguments{
\item{num_reps}{number of replicates}

\item{genes}{data frame of genes with the following specifications: first row is column labels, first column has gene labels/names, and all other columns have expression data. This expression data must be ordered by time point then by replicate, and must have evenly spaced time points. Any missing data must have cells left blank.}

\item{timen}{vector of time points for gene expression}
}
\value{
a matrix of means of gene expressions for replicates
}
\description{
Function to calculate the average of replicates for all given expressions. Used primarily for cases with multiple replicates.
}
