% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_package_functions.R
\name{calc_var}
\alias{calc_var}
\title{Function to calculate the variance of replicates at a certain time point.}
\usage{
calc_var(x, current_gene, num_reps, genes)
}
\arguments{
\item{x}{time point}

\item{current_gene}{row number of gene being examined}

\item{num_reps}{number of replicates}

\item{genes}{data frame of genes with the following specifications: first row is column labels, first column has gene labels/names, and all other columns have expression data. This expression data must be ordered by time point then by replicate, and must have evenly spaced time points. Any missing data must have cells left blank.}
}
\value{
the variance of replicates at a certain time point
}
\description{
Function to calculate the variance of replicates at a certain time point.
}
