% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_package_functions.R
\name{de_linear_trend_all}
\alias{de_linear_trend_all}
\title{Function to remove linear trend for each expression over all data.}
\usage{
de_linear_trend_all(timen, num_reps, tied, genes)
}
\arguments{
\item{timen}{vector of time points for all expressions}

\item{num_reps}{number of replicates}

\item{tied}{whether replicates are paired (tied) or unpaired (untied)}

\item{genes}{data frame of genes with the following specifications: first row is column labels, first column has gene labels/names, and all other columns have expression data. This expression data must be ordered by time point then by replicate, and must have evenly spaced time points. Any missing data must have cells left blank.}
}
\value{
a data frame containing the detrended expressions
}
\description{
Function to remove linear trend for each expression over all data.
}
