% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_find.R
\docType{package}
\name{echo.find}
\alias{echo.find}
\alias{echo.find-package}
\title{echo.find: Provides a function (echo_find) designed to find rhythms from data using extended harmonic oscillators.}
\description{
To read more about our inital work on this project and cite us, see Circadian Rhythms in Neurospora Exhibit Biologically Relevant Driven and Damped Harmonic Oscillations by H. De los Santos et al. (2017)
}
