% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_package_functions.R
\name{expansion.sum}
\alias{expansion.sum}
\title{Shewchuk algorithms for adaptive precision summation used in jtkdist
http://www.cs.cmu.edu/afs/cs/project/quake/public/papers/robust-arithmetic.ps}
\usage{
expansion.sum(g)
}
\arguments{
\item{g}{sum to expand}
}
\value{
two pieces of expanded sum
}
\description{
Shewchuk algorithms for adaptive precision summation used in jtkdist
http://www.cs.cmu.edu/afs/cs/project/quake/public/papers/robust-arithmetic.ps
}
\author{
Jonathan Shewchuk
}
