% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_package_functions.R
\name{fast.two.sum}
\alias{fast.two.sum}
\title{Shewchuk algorithms for adaptive precision summation used in jtkdist
http://www.cs.cmu.edu/afs/cs/project/quake/public/papers/robust-arithmetic.ps}
\usage{
fast.two.sum(a, b)
}
\arguments{
\item{a}{first number to sum}

\item{b}{first number to sum}
}
\value{
two different sums
}
\description{
Shewchuk algorithms for adaptive precision summation used in jtkdist
http://www.cs.cmu.edu/afs/cs/project/quake/public/papers/robust-arithmetic.ps
}
\author{
Jonathan Shewchuk
}
