% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_package_functions.R
\name{is_deviating}
\alias{is_deviating}
\title{Function for determining whether gene values are deviating or constant.}
\usage{
is_deviating(current_gene, genes)
}
\arguments{
\item{current_gene}{row number of current gene we want to calculate parameters for}

\item{genes}{data frame of genes with the following specifications: first row is column labels, first column has gene labels/names, and all other columns have expression data. This expression data must be ordered by time point then by replicate, and must have evenly spaced time points. Any missing data must have cells left blank.}
}
\value{
boolean if there is deviation within the gene values
}
\description{
Function for determining whether gene values are deviating or constant.
}
