% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_package_functions.R
\name{jtkdist}
\alias{jtkdist}
\title{details: http://www.jstor.org/pss/2347656}
\usage{
jtkdist(timepoints, reps = 1, normal = FALSE, alt = FALSE, jtk.alt)
}
\arguments{
\item{timepoints}{number of total amount of timepoints}

\item{reps}{number of replicates}

\item{normal}{boolean indicating whether a normal distribution is desired}

\item{alt}{boolean indicating whether there are missing time points}

\item{jtk.alt}{contains the exact p-value distribution for replicate data with missing data}
}
\value{
exact distribution of JTK, whether that be based on missing values or not
}
\description{
details: http://www.jstor.org/pss/2347656
}
\author{
E.F. Harding
}
