% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_package_functions.R
\name{normalize_all}
\alias{normalize_all}
\title{Function to normalize expressions in a matricized manner, by row.}
\usage{
normalize_all(genes)
}
\arguments{
\item{genes}{data frame of genes with the following specifications: first row is column labels, first column has gene labels/names, and all other columns have expression data. This expression data must be ordered by time point then by replicate, and must have evenly spaced time points. Any missing data must have cells left blank.}
}
\value{
normalized expression data
}
\description{
Function to normalize expressions in a matricized manner, by row.
}
